requirements.txt
Django==2.0.7
django-filter==2.0.0
djangorestframework==3.8.2
pytz==2018.5
